+++ src/utils/cookie.ts
@@ -1,7 +1,6 @@
 import { parse, serialize } from "cookie-es";
 import type { CookieSerializeOptions } from "cookie-es";
 import type { H3Event } from "../event";
-import { appendHeader } from "./response";
 
 /**
  * Parse the request to get HTTP Cookie header string and returning an object of all cookie name-value pairs.
@@ -48,7 +47,14 @@ export function setCookie(
     path: "/",
     ...serializeOptions,
   });
-  appendHeader(event, "Set-Cookie", cookieStr);
+  let setCookies = event.node.res.getHeader("set-cookie");
+  if (!Array.isArray(setCookies)) {
+    setCookies = [setCookies as any];
+  }
+  setCookies = setCookies.filter((cookieValue: string) => {
+    return cookieValue && !cookieValue.startsWith(name + "=");
+  });
+  event.node.res.setHeader("set-cookie", [...setCookies, cookieStr]);
 }
 
 /**