+++ src/utils/index.ts
@@ -6,3 +6,5 @@ export * from "./cookie";
 export * from "./proxy";
 export * from "./request";
 export * from "./response";
+export * from "./session";
+export * from "./cors";