+++ test/proxy.test.ts
@@ -11,6 +11,7 @@ import {
   getMethod,
   setHeader,
   readRawBody,
+  setCookie,
 } from "../src";
 import { sendProxy, proxyRequest } from "../src/utils/proxy";
 
@@ -83,11 +84,51 @@ describe("", () => {
       const result = await fetch(url + "/", {
         method: "POST",
         body: "hello",
-      }).then((r) => r.text());
+        headers: {
+          "content-type": "text/custom",
+          "x-custom": "hello",
+        },
+      }).then((r) => r.json());
 
-      expect(result).toMatchInlineSnapshot(
-        '"{\\"method\\":\\"POST\\",\\"headers\\":{\\"accept\\":\\"*/*\\",\\"accept-encoding\\":\\"gzip, deflate, br\\",\\"connection\\":\\"close\\",\\"content-length\\":\\"5\\",\\"content-type\\":\\"text/plain;charset=UTF-8\\",\\"user-agent\\":\\"node-fetch\\"},\\"body\\":\\"hello\\"}"'
+      const { headers, ...data } = result;
+      expect(headers["content-type"]).toEqual("text/custom");
+      expect(headers["x-custom"]).toEqual("hello");
+      expect(data).toMatchInlineSnapshot(`
+        {
+          "body": "hello",
+          "method": "POST",
+        }
+      `);
+    });
+  });
+
+  describe("multipleCookies", () => {
+    it("can split multiple cookies", async () => {
+      app.use(
+        "/setcookies",
+        eventHandler((event) => {
+          setCookie(event, "user", "alice", {
+            expires: new Date("Thu, 01 Jun 2023 10:00:00 GMT"),
+            httpOnly: true,
+          });
+          setCookie(event, "role", "guest");
+          return {};
+        })
       );
+
+      app.use(
+        "/",
+        eventHandler((event) => {
+          return sendProxy(event, url + "/setcookies", { fetch });
+        })
+      );
+
+      const result = await request.get("/");
+      const cookies = result.header["set-cookie"];
+      expect(cookies).toEqual([
+        "user=alice; Path=/; Expires=Thu, 01 Jun 2023 10:00:00 GMT; HttpOnly",
+        "role=guest; Path=/",
+      ]);
     });
   });